LPS-106031 Move openInfoItemSelector and reuse
